refactor: move constant value to a seperate folder
