+++ frontend/src/constants/appEnum.ts
